* lukas/gh/fix-release-sync-without-man-pages:
  gh: Fix build of missing release artifacts for 27
  gh: Update release sync script to work with 27 and patch release bodies

+++ .github/scripts/create-artifacts.sh
@@ -8,6 +8,7 @@ mkdir ${DIR}
 tar -xzf otp_src.tar.gz
 mv otp otp_src_${VSN}
 tar -czf ${DIR}/otp_src_${VSN}.tar.gz otp_src_${VSN}
-test -x otp_doc_man.tar.gz &&
+if [ -f otp_doc_man.tar.gz ]; then
     mv otp_doc_man.tar.gz ${DIR}/otp_doc_man_${VSN}.tar.gz
+fi
 mv otp_doc_html.tar.gz ${DIR}/otp_doc_html_${VSN}.tar.gz
